Move TLS settings into separate file
